Implementing CulturesSectionViewModel
